split api into 3 components
